+++ src/composio_client/types/org/__init__.py
@@ -7,5 +7,4 @@ from .project_list_response import ProjectListResponse as ProjectListResponse
 from .project_create_response import ProjectCreateResponse as ProjectCreateResponse
 from .project_delete_response import ProjectDeleteResponse as ProjectDeleteResponse
 from .api_key_retrieve_response import APIKeyRetrieveResponse as APIKeyRetrieveResponse
-from .project_retrieve_response import ProjectRetrieveResponse as ProjectRetrieveResponse
 from .api_key_regenerate_response import APIKeyRegenerateResponse as APIKeyRegenerateResponse